+++ apps/protocol-frontend/src/hooks/useContributionList.ts
@@ -2,6 +2,7 @@ import { useUser } from '../contexts/UserContext';
 import { useQuery, useInfiniteQuery } from '@tanstack/react-query';
 import { UIContribution } from '@govrn/ui-types';
 import { ListContributionsQueryVariables } from '@govrn/protocol-client';
+import { formatDate } from '../utils/date';
 
 export const useContributionList = (args?: ListContributionsQueryVariables) => {
   const { govrnProtocol: govrn } = useUser();
@@ -33,7 +34,15 @@ export const useContributionInfiniteList = (
       const data = await govrn.contribution.list(
         { ...args, first: pageSize, skip: pageParam } || {},
       );
-      return data.result;
+      const results = [];
+      for (const contribution of data.result) {
+        results.push({
+          ...contribution,
+          date_of_engagement: formatDate(contribution.date_of_engagement),
+          date_of_submission: formatDate(contribution.date_of_submission),
+        });
+      }
+      return results;
     },
     {
       getNextPageParam: (lastPage, pages) => {
