+++ apps/protocol-frontend/src/types/table.ts
@@ -20,3 +20,23 @@ export type ContributionTableType = {
   action: string;
   guildName: string;
 };
+
+export type AttestationTableType = {
+  id: number;
+  status: string;
+  contributor?: string | null;
+  date_of_submission: string | Date;
+  date_of_engagement: string | Date;
+  attestations?: {
+    id: number;
+  }[];
+  guilds?: {
+    guild: {
+      name?: string | null;
+      guild_id?: number;
+    };
+  }[];
+  action?: ReactNode;
+  name: string;
+  onChainId?: number | null;
+};
